requirements.txt
django
django-colorfield
django-extensions
